refacor: changed tree structure, added absolut paths and catch unknown commands and do not send them to the CoreMiner
