--- main.py
@@ -1,4 +1,0 @@
-from tui.app import HardHat
-
-if __name__ == "__main__":
-    HardHat().run()
