+++ src/css/main_view.tcss
