+++ src/css/widget_selector.tcss
