+++ src/widgets/display_response.py
@@ -3,4 +3,5 @@ from textual.widgets import Static
 class ResponseDisplay(Static):
     """Displays responses from CoreMiner"""
     def update_content(self, message: str):
-        self.update(f"Response: {message}")
+        self.update(f"Response: {message} ")
+    # Arguments do not get displayed at the moment
